+++ src/main/java/top/theillusivec4/curios/api/CuriosTriggers.java
@@ -21,10 +21,10 @@
 package top.theillusivec4.curios.api;
 
 import javax.annotation.Nonnull;
-import net.minecraft.advancements.Criterion;
 import net.minecraft.advancements.CriterionTriggerInstance;
-import net.minecraft.advancements.criterion.ItemPredicate;
-import net.minecraft.advancements.criterion.LocationPredicate;
+import net.minecraft.advancements.predicates.ItemPredicate;
+import net.minecraft.advancements.predicates.LocationPredicate;
+import net.minecraft.advancements.triggers.Criterion;
 
 public class CuriosTriggers {