+++ src/main/java/top/theillusivec4/curios/client/screen/CuriosScreenEvents.java
@@ -22,12 +22,12 @@ package top.theillusivec4.curios.client.screen;
 
 import com.mojang.blaze3d.platform.InputConstants;
 import com.mojang.blaze3d.platform.Window;
+import com.mojang.datafixers.util.Pair;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
 import net.minecraft.client.gui.screens.inventory.InventoryScreen;
-import net.minecraft.util.Tuple;
 import net.minecraft.world.inventory.Slot;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.neoforge.client.event.ScreenEvent;
@@ -50,13 +50,13 @@ public class CuriosScreenEvents {
     if (screen instanceof InventoryScreen || screen instanceof CreativeModeInventoryScreen) {
       AbstractContainerScreen<?> gui = (AbstractContainerScreen<?>) screen;
       boolean isCreative = screen instanceof CreativeModeInventoryScreen;
-      Tuple<Integer, Integer> offsets = CuriosScreen.getButtonOffset(isCreative);
-      int x = offsets.getA();
-      int y = offsets.getB();
+      Pair<Integer, Integer> offsets = CuriosScreen.getButtonOffset(isCreative);
+      int x = offsets.getFirst();
+      int y = offsets.getSecond();
       int size = isCreative ? 8 : 10;
       int yOffset = isCreative ? 67 : 81;
       evt.addListener(
-          new CuriosButton(gui, gui.getGuiLeft() + x - 2, gui.getGuiTop() + y + yOffset, size, size,
+          new CuriosButton(gui, gui.getLeftPos() + x - 2, gui.getTopPos() + y + yOffset, size, size,
                            isCreative ? CuriosButton.SMALL : CuriosButton.BIG));
     }
   }
@@ -73,7 +73,7 @@ public class CuriosScreenEvents {
       return;
     }
     Slot destroyItemSlot = gui.destroyItemSlot;
-    Slot slot = gui.getSlotUnderMouse();
+    Slot slot = gui.getHoveredSlot();
 
     if (destroyItemSlot != null && slot == destroyItemSlot) {
       ClientPacketDistributor.sendToServer(new CPacketDestroy());