+++ src/main/java/top/theillusivec4/curios/common/CuriosCommonEvents.java
@@ -43,7 +43,6 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.players.PlayerList;
 import net.minecraft.util.Mth;
-import net.minecraft.util.Tuple;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
@@ -104,7 +103,7 @@ import top.theillusivec4.curios.impl.CuriosRegistry;
 public class CuriosCommonEvents {
 
   private static void handleDrops(String identifier, LivingEntity livingEntity,
-                                  List<Tuple<Predicate<ItemStack>, DropRule>> dropRules,
+                                  List<Pair<Predicate<ItemStack>, DropRule>> dropRules,
                                   NonNullList<Boolean> renders, IDynamicStackHandler stacks,
                                   boolean cosmetic, Collection<ItemEntity> drops,
                                   boolean keepInventory, LivingDropsEvent evt) {
@@ -116,10 +115,10 @@ public class CuriosCommonEvents {
       if (!stack.isEmpty()) {
         DropRule dropRuleOverride = null;
 
-        for (Tuple<Predicate<ItemStack>, DropRule> override : dropRules) {
+        for (Pair<Predicate<ItemStack>, DropRule> override : dropRules) {
 
-          if (override.getA().test(stack)) {
-            dropRuleOverride = override.getB();
+          if (override.getFirst().test(stack)) {
+            dropRuleOverride = override.getSecond();
           }
         }
         DropRule dropRule = dropRuleOverride != null ? dropRuleOverride : CuriosApi.getCurio(stack)
@@ -276,7 +275,7 @@ public class CuriosCommonEvents {
         DropRulesEvent dropRulesEvent =
             new DropRulesEvent(livingEntity, handler, evt.getSource(), 0, evt.isRecentlyHit());
         NeoForge.EVENT_BUS.post(dropRulesEvent);
-        List<Tuple<Predicate<ItemStack>, DropRule>> dropRules = dropRulesEvent.getOverrides();
+        List<Pair<Predicate<ItemStack>, DropRule>> dropRules = dropRulesEvent.getOverrides();
         boolean keepInventory = false;
 
         if (livingEntity instanceof Player
@@ -332,7 +331,7 @@ public class CuriosCommonEvents {
     CuriosApi.getCurio(stack).ifPresent(
         curio -> CuriosApi.getCuriosInventory(player).ifPresent(handler -> {
           Map<String, ICurioStacksHandler> curios = handler.getCurios();
-          Tuple<IDynamicStackHandler, SlotContext> firstSlot = null;
+          Pair<IDynamicStackHandler, SlotContext> firstSlot = null;
 
           for (Map.Entry<String, ICurioStacksHandler> entry : curios.entrySet()) {
             IDynamicStackHandler stackHandler = entry.getValue().getStacks();
@@ -370,7 +369,7 @@ public class CuriosCommonEvents {
 
                   if (stackHandler.extractItem(i, stack.getMaxStackSize(), true).getCount() ==
                       stack.getCount()) {
-                    firstSlot = new Tuple<>(stackHandler, slotContext);
+                    firstSlot = new Pair<>(stackHandler, slotContext);
                   }
                 }
               }
@@ -378,8 +377,8 @@ public class CuriosCommonEvents {
           }
 
           if (firstSlot != null) {
-            IDynamicStackHandler stackHandler = firstSlot.getA();
-            SlotContext slotContext = firstSlot.getB();
+            IDynamicStackHandler stackHandler = firstSlot.getFirst();
+            SlotContext slotContext = firstSlot.getSecond();
             int i = slotContext.index();
             ItemStack present = stackHandler.getStackInSlot(i);
             stackHandler.setStackInSlot(i, stack.copy());