+++ src/main/java/top/theillusivec4/curios/impl/CuriosRegistry.java
@@ -21,7 +21,7 @@
 package top.theillusivec4.curios.impl;
 
 import java.util.function.Supplier;
-import net.minecraft.advancements.CriterionTrigger;
+import net.minecraft.advancements.triggers.CriterionTrigger;
 import net.minecraft.commands.synchronization.ArgumentTypeInfo;
 import net.minecraft.commands.synchronization.ArgumentTypeInfos;
 import net.minecraft.commands.synchronization.SingletonArgumentInfo;