+++ src/main/java/top/theillusivec4/curios/mixin/core/MixinNbtPredicate.java
@@ -20,7 +20,7 @@
 
 package top.theillusivec4.curios.mixin.core;
 
-import net.minecraft.advancements.criterion.NbtPredicate;
+import net.minecraft.advancements.predicates.NbtPredicate;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.util.ProblemReporter;
 import net.minecraft.world.entity.Entity;