+++ src/test/java/top/theillusivec4/curiostest/data/CuriosTestProvider.java
@@ -3,14 +3,14 @@ package top.theillusivec4.curiostest.data;
 import java.util.concurrent.CompletableFuture;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
-import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.item.Items;
+import net.minecraft.references.ItemIds;
+import net.minecraft.world.entity.EntityTypes;
 import net.neoforged.neoforge.common.conditions.NeoForgeConditions;
 import top.theillusivec4.curios.api.CuriosDataProvider;
 import top.theillusivec4.curios.api.CuriosTags;
 import top.theillusivec4.curios.api.common.DropRule;
 import top.theillusivec4.curios.api.type.data.ISlotData;
-import top.theillusivec4.curiostest.common.CuriosTestRegistry;
+import top.theillusivec4.curiostest.common.CuriosTestIds;
 
 public class CuriosTestProvider extends CuriosDataProvider {
 
@@ -31,16 +31,16 @@ public class CuriosTestProvider extends CuriosDataProvider {
         .addCondition(NeoForgeConditions.never());
 
     ISlotData testSlot = this.createSlot("slot_from_slots").size(2);
-    this.tag(testSlot).add(Items.DIAMOND);
+    this.tag(testSlot).add(ItemIds.DIAMOND);
 
-    this.tag(CuriosTags.HANDS).add(CuriosTestRegistry.KNUCKLES.get());
-    this.tag(CuriosTags.NECKLACE).add(CuriosTestRegistry.AMULET.get());
-    this.tag(CuriosTags.RING).add(CuriosTestRegistry.RING.get());
-    this.tag(CuriosTags.HEAD).add(CuriosTestRegistry.CROWN.get());
+    this.tag(CuriosTags.HANDS).add(CuriosTestIds.KNUCKLES);
+    this.tag(CuriosTags.NECKLACE).add(CuriosTestIds.AMULET);
+    this.tag(CuriosTags.RING).add(CuriosTestIds.RING);
+    this.tag(CuriosTags.HEAD).add(CuriosTestIds.CROWN);
 
     this.createEntities("test_entities")
         .addPlayer()
-        .addEntities(EntityType.WITHER_SKELETON)
+        .addEntities(EntityTypes.WITHER_SKELETON)
         .addAllPresetSlots()
         .addSlots("inline_from_entities")
         .addSlots(this.getSlot("slot_from_entities").size(3))