+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html). Prior to version 5.2.0, this projected
 used [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [16.0.0+26.2] - 2026.07.20
+
+### Changed
+- [API] `top.theillusivec4.curios.api.event.DropRulesEvent#getOverrides()` refactored `net.minecraft.util.Tuple`
+  return type to `com.mojang.datafixers.util.Pair`
+- Cleaned up mixin debug logs
+- Updated to Minecraft 26.2
+
+### Removed
+- [API] Removed deprecated methods in `top.theillusivec4.curios.api.client.ICurioRenderer` that took a
+  `MultiBufferSource` parameter:
+  - `render` (use the `SubmitNodeCollector` replacement method)
+  - `renderModel`
+
 ## [15.0.0+26.1.2] - 2026.07.19
 
 ### Added